+++ modules/apps/document-library/document-library-service/src/main/java/com/liferay/document/library/internal/trash/DLFileShortcutTrashRendererFactory.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.document.library.web.internal.trash;
+package com.liferay.document.library.internal.trash;
 
 import com.liferay.document.library.kernel.model.DLFileShortcut;
 import com.liferay.document.library.kernel.service.DLFileShortcutLocalService;
@@ -20,7 +20,6 @@ import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.trash.TrashRenderer;
 import com.liferay.portal.kernel.trash.TrashRendererFactory;
 import com.liferay.trash.TrashHelper;
-
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
