+++ modules/apps/message-boards/message-boards-service/src/main/java/com/liferay/message/boards/internal/trash/MBCategoryTrashRenderer.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.message.boards.web.internal.trash;
+package com.liferay.message.boards.internal.trash;
 
 import com.liferay.asset.kernel.AssetRendererFactoryRegistryUtil;
 import com.liferay.asset.kernel.model.AssetRendererFactory;
