+++ modules/apps/message-boards/message-boards-service/src/main/java/com/liferay/message/boards/internal/trash/MBThreadTrashRenderer.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.message.boards.web.internal.trash;
+package com.liferay.message.boards.internal.trash;
 
 import com.liferay.asset.constants.AssetWebKeys;
 import com.liferay.asset.util.AssetHelper;
